feat(dashboard): promote root keys to top-level workspace navigation
Root Keys gets its own /root-keys route below Projects in the
sidebar; the settings URL redirects there and leaves the settings
secondary nav.
